splitup mapset_management
